refactor rendering for prerender
